requirements.txt
fastapi
databricks-sdk
reportlab
requests
uvicorn
